+++ jason/CHANGELOG.md
@@ -26,7 +26,7 @@ All user visible changes to this project will be documented in this file. This p
 
 - Media management:
     - Library API:
-        - Mute/unmute local video/audio ([#40], [#81], [#97]):
+        - Mute/unmute local video/audio ([#40], [#81], [#97], [#144]):
             - `Room.mute_audio()`;
             - `Room.unmute_audio()`;
             - `Room.mute_video()`;
@@ -57,6 +57,7 @@ All user visible changes to this project will be documented in this file. This p
             - `Room.unmute_remote_video`.
         - `MediaTrack.media_source_kind` function ([#145], [#146]).
     - Optional tracks support ([#106]);
+    - Simultaneous device and display video tracks publishing and receiving ([#144]);
     - `RtcIceTransportPolicy` configuration ([#79]).
 - Room management:
     - Library API:
@@ -108,6 +109,7 @@ All user visible changes to this project will be documented in this file. This p
 [#137]: /../../pull/137
 [#138]: /../../pull/138
 [#143]: /../../pull/143
+[#144]: /../../pull/144
 [#145]: /../../pull/145
 [#146]: /../../pull/146